+++ src/abstractdataframe/sort.jl
@@ -257,7 +257,7 @@ function ordering(df::AbstractDataFrame, cols::AbstractVector, lt::AbstractVecto
     end
 
     if length(cols) == 1
-        return ordering(df, cols[1], lt[1], by[1], rev[1], order[1])
+        return ordering(df, only(cols), only(lt), only(by), only(rev), only(order))
     end
 
     # Collect per-column ordering info
@@ -333,31 +333,45 @@ Sort.defalg(df::AbstractDataFrame, o::Ordering; alg=nothing, cols=[]) =
 ## Actual sort functions
 ########################
 
+const SORT_ARGUMENTS =
 """
-    issorted(df::AbstractDataFrame, cols=All();
-             lt=isless, by=identity, rev::Bool=false, order::Ordering=Forward)
-
-Test whether data frame `df` sorted by column(s) `cols`.
-Checking against multiple columns is done lexicographically.
-
-`cols` can be any column selector ($COLUMNINDEX_STR; $MULTICOLUMNINDEX_STR).
-If `cols` selects no columns, check whether `df` is sorted on all columns
-(this behaviour is deprecated and will change in future versions).
-
-`cols` can be any column selector ($COLUMNINDEX_STR; $MULTICOLUMNINDEX_STR).
-
-If `rev` is `true`, reverse sorting is performed. To enable reverse sorting
-only for some columns, pass `order(c, rev=true)` in `cols`, with `c` the
+If `rev` is `true`, reverse sorting is performed. To enable reverse sorting only
+for some columns, pass `order(c, rev=true)` in `cols`, with `c` the
 corresponding column index (see example below).
 
-See other methods for a description of other keyword arguments.
+The `by` keyword allows providing a function that will be applied to each
+cell before comparison; the `lt` keyword allows providing a custom "less
+than" function. If both `by` and `lt` are specified, the `lt` function is
+applied to the result of the `by` function.
+
+All the keyword arguments can be either a single value, which is applied to
+all columns, or a vector of length equal to the number of columns that the
+operation is performed on. In such a case each entry is used for the
+column in the corresponding position in `cols`.
+"""
+
+"""
+    issorted(df::AbstractDataFrame, cols=All();
+             lt::Union{Function, AbstractVector{<:Function}}=isless,
+             by::Union{Function, AbstractVector{<:Function}}=identity,
+             rev::Union{Bool, AbstractVector{Bool}}=false,
+             order::Union{Ordering, AbstractVector{<:Ordering}}=Forward)
+
+Test whether data frame `df` sorted by column(s) `cols`. Checking against
+multiple columns is done lexicographically.
+
+`cols` can be any column selector ($COLUMNINDEX_STR; $MULTICOLUMNINDEX_STR). If
+`cols` selects no columns, check whether `df` is sorted on all columns (this
+behaviour is deprecated and will change in future versions).
+
+$SORT_ARGUMENTS
 
 # Examples
 ```jldoctest
 julia> df = DataFrame(a = [1, 2, 3, 4], b = [4, 3, 2, 1])
 4×2 DataFrame
- Row │ a      b     
-     │ Int64  Int64 
+ Row │ a      b
+     │ Int64  Int64
 ─────┼──────────────
    1 │     1      4
    2 │     2      3
@@ -378,15 +392,23 @@ true
 ```
 """
 function Base.issorted(df::AbstractDataFrame, cols=All();
-                       lt=isless, by=identity, rev=false, order=Forward)
+                       lt::Union{Function, AbstractVector{<:Function}}=isless,
+                       by::Union{Function, AbstractVector{<:Function}}=identity,
+                       rev::Union{Bool, AbstractVector{Bool}}=false,
+                       order::Union{Ordering, AbstractVector{<:Ordering}}=Forward)
+    to_scalar(x::AbstractVector) = only(x)
+    to_scalar(x::Any) = x
+
     # exclude AbstractVector as in that case cols can contain order(...) clauses
     if cols isa MultiColumnIndex && !(cols isa AbstractVector)
         cols = index(df)[cols]
     end
     if cols isa ColumnIndex
-        return issorted(df[!, cols], lt=lt, by=by, rev=rev, order=order)
+        return issorted(df[!, cols], lt=to_scalar(lt), by=to_scalar(by),
+                        rev=to_scalar(rev), order=to_scalar(order))
     elseif cols isa AbstractVector{<:ColumnIndex} && length(cols) == 1
-        return issorted(df[!, cols[1]], lt=lt, by=by, rev=rev, order=order)
+        return issorted(df[!, cols[1]], lt=to_scalar(lt), by=to_scalar(by),
+                        rev=to_scalar(rev), order=to_scalar(order))
     else
         return issorted(1:nrow(df), ordering(df, cols, lt, by, rev, order))
     end
@@ -394,8 +416,12 @@ end
 
 """
     sort(df::AbstractDataFrame, cols=All();
-         alg::Union{Algorithm, Nothing}=nothing, lt=isless, by=identity,
-         rev::Bool=false, order::Ordering=Forward, view::Bool=false)
+         alg::Union{Algorithm, Nothing}=nothing,
+         lt::Union{Function, AbstractVector{<:Function}}=isless,
+         by::Union{Function, AbstractVector{<:Function}}=identity,
+         rev::Union{Bool, AbstractVector{Bool}}=false,
+         order::Union{Ordering, AbstractVector{<:Ordering}}=Forward,
+         view::Bool=false)
 
 Return a data frame containing the rows in `df` sorted by column(s) `cols`.
 Sorting on multiple columns is done lexicographically.
@@ -404,18 +430,14 @@ Sorting on multiple columns is done lexicographically.
 If `cols` selects no columns, sort `df` on all columns
 (this behaviour is deprecated and will change in future versions).
 
+$SORT_ARGUMENTS
+
 If `alg` is `nothing` (the default), the most appropriate algorithm is
 chosen automatically among `TimSort`, `MergeSort` and `RadixSort` depending
 on the type of the sorting columns and on the number of rows in `df`.
 
-If `rev` is `true`, reverse sorting is performed. To enable reverse sorting
-only for some columns, pass `order(c, rev=true)` in `cols`, with `c` the
-corresponding column index (see example below).
-
 If `view=false` a freshly allocated `DataFrame` is returned.
 If `view=true` then a `SubDataFrame` view into `df` is returned.
-
-See [`sort!`](@ref) for a description of other keyword arguments.
 
 # Examples
 ```jldoctest
@@ -470,16 +492,24 @@ julia> sort(df, [:x, order(:y, rev=true)])
    4 │     3  b
 ```
 """
-@inline function Base.sort(df::AbstractDataFrame, cols=All(); alg=nothing, lt=isless,
-                           by=identity, rev=false, order=Forward, view::Bool=false)
+@inline function Base.sort(df::AbstractDataFrame, cols=All();
+                           alg::Union{Algorithm, Nothing}=nothing,
+                           lt::Union{Function, AbstractVector{<:Function}}=isless,
+                           by::Union{Function, AbstractVector{<:Function}}=identity,
+                           rev::Union{Bool, AbstractVector{Bool}}=false,
+                           order::Union{Ordering, AbstractVector{<:Ordering}}=Forward,
+                           view::Bool=false)
     rowidxs = sortperm(df, cols, alg=alg, lt=lt, by=by, rev=rev, order=order)
     return view ? Base.view(df, rowidxs, :) : df[rowidxs, :]
 end
 
 """
     sortperm(df::AbstractDataFrame, cols=All();
-             alg::Union{Algorithm, Nothing}=nothing, lt=isless, by=identity,
-             rev::Bool=false, order::Ordering=Forward)
+             alg::Union{Algorithm, Nothing}=nothing,
+             lt::Union{Function, AbstractVector{<:Function}}=isless,
+             by::Union{Function, AbstractVector{<:Function}}=identity,
+             rev::Union{Bool, AbstractVector{Bool}}=false,
+             order::Union{Ordering, AbstractVector{<:Ordering}}=Forward)
 
 Return a permutation vector of row indices of data frame `df` that puts them in
 sorted order according to column(s) `cols`.
@@ -489,15 +519,11 @@ Order on multiple columns is computed lexicographically.
 If `cols` selects no columns, return permutation vector based on sorting all columns
 (this behaviour is deprecated and will change in future versions).
 
+$SORT_ARGUMENTS
+
 If `alg` is `nothing` (the default), the most appropriate algorithm is
 chosen automatically among `TimSort`, `MergeSort` and `RadixSort` depending
 on the type of the sorting columns and on the number of rows in `df`.
-
-If `rev` is `true`, reverse sorting is performed. To enable reverse sorting
-only for some columns, pass `order(c, rev=true)` in `cols`, with `c` the
-corresponding column index (see example below).
-
-See other methods for a description of other keyword arguments.
 
 # Examples
 ```jldoctest
@@ -541,12 +567,11 @@ julia> sortperm(df, [:x, order(:y, rev=true)])
 ```
 """
 function Base.sortperm(df::AbstractDataFrame, cols=All();
-                  alg=nothing, lt=isless, by=identity, rev=false, order=Forward)
-    if !(by isa Base.Callable || (by isa AbstractVector && eltype(by) <: Base.Callable))
-        msg = "'by' must be a Function or a vector of Functions. " *
-              " Perhaps you wanted 'cols'."
-        throw(ArgumentError(msg))
-    end
+                       alg::Union{Algorithm, Nothing}=nothing,
+                       lt::Union{Function, AbstractVector{<:Function}}=isless,
+                       by::Union{Function, AbstractVector{<:Function}}=identity,
+                       rev::Union{Bool, AbstractVector{Bool}}=false,
+                       order::Union{Ordering, AbstractVector{<:Ordering}}=Forward)
     # exclude AbstractVector as in that case cols can contain order(...) clauses
     if cols isa MultiColumnIndex && !(cols isa AbstractVector)
         cols = index(df)[cols]
@@ -560,3 +585,119 @@ _sortperm(df::AbstractDataFrame, a::Algorithm, o::Union{Perm, DFPerm}) =
     sort!([1:size(df, 1);], a, o)
 _sortperm(df::AbstractDataFrame, a::Algorithm, o::Ordering) =
     sortperm(df, a, DFPerm(o, df))
+
+
+"""
+    sort!(df::AbstractDataFrame, cols=All();
+          alg::Union{Algorithm, Nothing}=nothing,
+          lt::Union{Function, AbstractVector{<:Function}}=isless,
+          by::Union{Function, AbstractVector{<:Function}}=identity,
+          rev::Union{Bool, AbstractVector{Bool}}=false,
+          order::Union{Ordering, AbstractVector{<:Ordering}}=Forward)
+
+Sort data frame `df` by column(s) `cols`.
+Sorting on multiple columns is done lexicographicallly.
+
+`cols` can be any column selector ($COLUMNINDEX_STR; $MULTICOLUMNINDEX_STR). If
+`cols` selects no columns, sort `df` on all columns (this
+behaviour is deprecated and will change in future versions).
+
+$SORT_ARGUMENTS
+
+If `alg` is `nothing` (the default), the most appropriate algorithm is
+chosen automatically among `TimSort`, `MergeSort` and `RadixSort` depending
+on the type of the sorting columns and on the number of rows in `df`.
+
+# Examples
+```jldoctest
+julia> df = DataFrame(x = [3, 1, 2, 1], y = ["b", "c", "a", "b"])
+4×2 DataFrame
+ Row │ x      y
+     │ Int64  String
+─────┼───────────────
+   1 │     3  b
+   2 │     1  c
+   3 │     2  a
+   4 │     1  b
+
+julia> sort!(df, :x)
+4×2 DataFrame
+ Row │ x      y
+     │ Int64  String
+─────┼───────────────
+   1 │     1  c
+   2 │     1  b
+   3 │     2  a
+   4 │     3  b
+
+julia> sort!(df, [:x, :y])
+4×2 DataFrame
+ Row │ x      y
+     │ Int64  String
+─────┼───────────────
+   1 │     1  b
+   2 │     1  c
+   3 │     2  a
+   4 │     3  b
+
+julia> sort!(df, [:x, :y], rev=true)
+4×2 DataFrame
+ Row │ x      y
+     │ Int64  String
+─────┼───────────────
+   1 │     3  b
+   2 │     2  a
+   3 │     1  c
+   4 │     1  b
+
+julia> sort!(df, [:x, order(:y, rev=true)])
+4×2 DataFrame
+ Row │ x      y
+     │ Int64  String
+─────┼───────────────
+   1 │     1  c
+   2 │     1  b
+   3 │     2  a
+   4 │     3  b
+```
+"""
+function Base.sort!(df::AbstractDataFrame, cols=All();
+                    alg::Union{Algorithm, Nothing}=nothing,
+                    lt::Union{Function, AbstractVector{<:Function}}=isless,
+                    by::Union{Function, AbstractVector{<:Function}}=identity,
+                    rev::Union{Bool, AbstractVector{Bool}}=false,
+                    order::Union{Ordering, AbstractVector{<:Ordering}}=Forward)
+
+    # exclude AbstractVector as in that case cols can contain order(...) clauses
+    if cols isa MultiColumnIndex && !(cols isa AbstractVector)
+        cols = index(df)[cols]
+    end
+    ord = ordering(df, cols, lt, by, rev, order)
+    _alg = Sort.defalg(df, ord; alg=alg, cols=cols)
+    return sort!(df, _alg, ord)
+end
+
+function Base.sort!(df::AbstractDataFrame, a::Base.Sort.Algorithm, o::Base.Sort.Ordering)
+    c = collect(eachcol(df))
+
+    toskip = Set{Int}()
+    for (i, col) in enumerate(c)
+        # Check if this column has been sorted already
+        if any(j -> c[j] === col, 1:i-1)
+            push!(toskip, i)
+        elseif any(j -> Base.mightalias(c[j], col), 1:i-1)
+            throw(ArgumentError("data frame contains non identical columns that share the same memory"))
+        end
+    end
+
+    p = _sortperm(df, a, o)
+    pp = similar(p)
+
+    for (i, col) in enumerate(c)
+        if !(i in toskip)
+            copyto!(pp, p)
+            Base.permute!!(col, pp)
+        end
+    end
+    return df
+end