+++ NEWS.md
@@ -92,6 +92,8 @@
   to compute standard deviation and quartiles
   in addition to statistics that are reported by default.
   ([#2459](https://github.com/JuliaData/DataFrames.jl/pull/2459))
+* `sort!` now supports general `AbstractDataFrame`
+  ([#2946](https://github.com/JuliaData/DataFrames.jl/pull/2946))
 
 ## Bug fixes
 
@@ -99,6 +101,10 @@
   ([#2842](https://github.com/JuliaData/DataFrames.jl/issues/2842))
 * fix a problem with not specialized `Pair` arguments passed as transformations
   ([#2889](https://github.com/JuliaData/DataFrames.jl/issues/2889))
+* sorting related functions now more carefully check passed arguments for
+  correctness. Now all keyword arguments are correctly checked to be either
+  scalars of vectors of scalars.
+  ([#2946](https://github.com/JuliaData/DataFrames.jl/pull/2946))
 
 ## Performance improvements